Rename passwords -> secrets
I'm not just using this directory to store passwords, so don't imply
that I am.

Add a check for the old directory being present, so I can remove it if
it is.

Also move user configuration to a NixOS module, so it's easier to add
more complex user configuration, including importing the user password
file from wherever it might be.