Overlays: rearrange significantly
Emulate the Nixpkgs by-name directory handling in the pkgs.nix overlay,
and similarly simplify pkgs/default.nix.

Add the overlay directory to the user environment as part of building a
Home Manager environment.

Always use all overlays present in the local directory when building a
Home Manager environment or NixOS system.  This makes things that rely
on the overlays slightly less standalone -- dependencies on overlays
will be less obvious -- but it means overlays only need to be applied
once, rather than once for every module or package that might use one.