NixOS/Nix: fail quickly when substituters fail
If I have multiple substituters, I want to use later ones even if the
earlier ones aren't available.  In particular, if my laptop is
configured with a binary cache on another system in my local network, I
want it to use that cache when it can, but to use the public cache when
I'm not at home.

Set up Nix to fail more quickly when a substituter isn't available, and
to keep trying later substituters when earlier ones aren't available
rather than giving up entirely.