Home/Homeshick: use activation script for activation
When I initially wrote the Homeshick Home Manager module, I stuck with
what I was familiar with: I wrote the code for setting up the Homeshick
repositories using Systemd units.  Now I know better!  Do the same
thing, but using the Home Manager activation script.

While we're here, change the meaning of the forceLink configuration
option to be slightly more intuitive: it's ignored if the link
attribute isn't true.  And also remove the dependency on GitHub CLI;
that might be useful for me sometimes, but it's not something that's
generally useful.  Provide a config toggle to add extra runtime
dependencies during the activation phase to include such tools.