taskloop: option to ignore SIGWINCH
If I'm using tmux-taskloop, that will handle SIGWINCH for me more
sensibly than trying to catch it in my Bash script, so provide an option
to ignore it in Bash.

While we're here, improve the option handling to be able to cope with
multiple switches in a single argument, and tidy the tasklooprc path
specification.