OpenVPN: make systemd notification happen later
Upstream OpenVPN tells systemd that it's ready as soon as its initial
setup is complete, which is notably before the VPN connection is ready
to use.  I want to order units after the VPN systemd unit, which means I
want the "ready" notification to not happen until the VPN is actually
ready to be used.  To make that happen, overlay a patch on the upstream
OpenVPN code.