NixOS/OpenVPN: don't error after sleep
OpenVPN NixOS config has a small unit that restarts OpenVPN clients
that were running when the system went to sleep.  Which is fine and
reasonable, except it produces errors if there weren't any clients
running before the system went to sleep.  I've submitted a patch
upstream for this, but in the meantime apply that same change locally.