NixOS/Scan to OneDrive: remove home directory
The scanner user doesn't need a home directory, so don't create one.
Definitely don't create one that's also the upload directory, as that
seems to cause NixOS's user management to force it to only be readable
by the owner, where I want it to be both readable and writable by the
group.