Use absolute paths
Rather than using relative paths for relative files, in particular
including the secrets directory, use absolute paths.  In particular,
this means building this configuration using nixos-install in /mnt or
similar works, rather than having the new installation pointing to a
bunch of paths that reference a mountpoint that has moved when booting
into the system itself.