git-tip: rewrite to better use upstream builds
Rather than having a bunch of my own custom code in the git-tip build
process, rewrite the source package handling to use a source bundle that
looks much more like the source bundles the Git project provides and
that the Nixpkgs builds expect.

While we're at it, rewrite the code for getting the source code to be
able to cope with tags as well as branches.