taskloop: remove standalone package
I'm only using the taskloop script as part of tmux-taskloop these days,
so remove the standalone package and just make it part of tmux-taskloop.
This will make it easier to change the taskloop script behaviour to
better integrate with tmux-taskloop, rather than taskloop needing to
work as a standalone script as well.