Flake/modules: rearrange imports
Rather than working out the other files to import as part of the
default.nix module, work out all the modules to import at the flake
level.  This means it's possible to use the flake's lib.dirfiles without
worrying about what inputs are exposed to the module or the special
early-in-module-processing handling that the imports attribute receives.