Flake: create an "everything" attribute
This is a new attribute that can't easily be built anywhere as it
inherently requires outputs from all architectures I'm building with.
What it does do, however, is depend on every derivation for all
architectures under the "checks" attribute.  If I work out the
derivation path for this attribute, it will let me see what difference
some change makes across all my systems, configurations and packages
with nix-diff or similar.