pkgs: move definition
There's a problem with defining my set of packages in the flake, then
referencing that in the overlay: it means the packages added by the
overlay are forced to use the version of nixpkgs instantiated in the
flake, rather than the version that comes from the `final` argument,
which may have different configuration or other relevant overlays
included.

Instead, define my packages in the overlay, then reference that
definition in the definition of legacyPackages in the flake.